begin the recurseInverse function started with B and D

+++ matrix.cpp
@@ -14,9 +14,8 @@ class Matrix{
         long unsigned rows;
         long unsigned cols;
 
-
         public:
-            int ** arr;
+            double ** arr;
 
 
     public:
